layout multistep dialog:wq
