+++ _inc/client/components/jetpack-disconnect-dialog-v2/survey.jsx
@@ -1,0 +1,9 @@
+/**
+ * External dependencies
+ */
+import React from 'react';
+
+const JetpackDisconnectDialogSurvey = ( { children } ) => {
+	return <div>{ children }</div>;
+};
+export default JetpackDisconnectDialogSurvey;
